+++ services/data_plane_manager/src/main/java/com/futurewei/alcor/dataplane/service/ovs/DataPlaneServiceImplNew.java
@@ -88,7 +88,7 @@ public class DataPlaneServiceImplNew implements DataPlaneServiceNew {
             vpcConfigBuilder.setProjectId(vpcEntity.getProjectId());
             vpcConfigBuilder.setName(vpcEntity.getName());
             vpcConfigBuilder.setCidr(vpcEntity.getCidr());
-            vpcConfigBuilder.setTunnelId(Long.parseLong(vpcEntity.getSegmentationId() + ""));
+            vpcConfigBuilder.setTunnelId(Integer.parseInt(vpcEntity.getSegmentationId() + ""));
 
             networkConfig.getSubnets().stream()
                     .filter(s -> s.getVpcId().equals(vpcEntity.getId()))
